Add io-uring madvise page loading on Linux

+++ src/main.rs
@@ -9,7 +9,7 @@ fn main() {
     let vec: Vec<_> = (0..10_000_000).collect();
     let value = std::env::args()
         .nth(1)
-        .and_then(|s| s.parse::<i32>().ok())
+        .and_then(|s| s.parse::<i64>().ok())
         .unwrap_or(10_000);
 
     let bsa = binary_search_cor(vec.as_slice(), value);
